Make funtion templates real functions
Prompts remplates are currently contained in the docstring of decorated
functions. The main issue with this is that prompt templates cannot be
composed.

In this commit we instead require users to return the prompt
template from the function. The template will then automatically be
rendered using the values passed to the function. This is very flexible:
some variables can be used inside the functions and not be present in
the Jinja2 template that is returned, for instance:

```python
import prompts

@prompts.template
def my_template(a, b):
    prompt = f'This is a first variable {a}'
    return prompt + "and a second {{b}}"
```